+++ src/main/java/com/cognixia/jump/model/CheckingsAccount.java
@@ -17,12 +17,12 @@ public class CheckingsAccount extends Account{
 
 	
 	@Override
-	public void deposit(int amount) {
+	public void deposit(Double amount) {
 		this.balance += amount;
 	}
 
 	@Override
-	public void withdraw(int amount) {
+	public void withdraw(Double amount) {
 		this.balance -= amount;
 	}
